+++ auth/auth.go
@@ -1,35 +1,33 @@
 package auth
 
-import (
-	"github.com/mainflux/fluxmq/pkg/client"
-)
+import "github.com/mainflux/fluxmq/client"
 
-// Handler is an interface for FluxMQ hooks
+// AuthHandler is an interface for FluxMQ hooks
 type Handler interface {
 	// Authorization on client `CONNECT`
 	// Each of the params are passed by reference, so that it can be changed
-	AuthConnect(client *client.ClientInfo) error
+	AuthConnect(client *client.Info) error
 
 	// Authorization on client `PUBLISH`
 	// Topic is passed by reference, so that it can be modified
-	AuthPublish(client *client.ClientInfo, topic *string, payload *[]byte) error
+	AuthPublish(client *client.Info, topic *string, payload *[]byte) error
 
 	// Authorization on client `SUBSCRIBE`
 	// Topics are passed by reference, so that they can be modified
-	AuthSubscribe(client *client.ClientInfo, topics *[]string) error
+	AuthSubscribe(client *client.Info, topic *string) error
 
 	// After client successfully connected
-	Connect(client *client.ClientInfo)
+	Connect(client *client.Info)
 
 	// After client successfully published
-	Publish(client *client.ClientInfo, topic *string, payload *[]byte)
+	Publish(client *client.Info, topic *string, payload *[]byte)
 
 	// After client successfully subscribed
-	Subscribe(client *client.ClientInfo, topics *[]string)
+	Subscribe(client *client.Info, topics *[]string)
 
 	// After client unsubscribed
-	Unsubscribe(client *client.ClientInfo, topics *[]string)
+	Unsubscribe(client *client.Info, topics *[]string)
 
 	// Disconnect on connection with client lost
-	Disconnect(client *client.ClientInfo)
+	Disconnect(client *client.Info)
 }
