+++ session/repository.go
@@ -13,3 +13,8 @@ func NewRepository() *Repository {
 		Sessions: make(map[string]Session),
 	}
 }
+
+func (r *Repository) Session(clientID string) (Session, bool) {
+	val, ok := r.Sessions[clientID]
+	return val, ok
+}
